64-bit changes: ioctl cmd -> u_long, some protos.  some style, return vals.
SCCS-vsn: sys/sys/socketvar.h 8.2
SCCS-vsn: sys/sys/proc.h 8.13
SCCS-vsn: sys/sys/acct.h 8.4
SCCS-vsn: sys/sys/ucred.h 8.4
SCCS-vsn: sys/sys/mbuf.h 8.4
SCCS-vsn: sys/sys/disk.h 8.2
SCCS-vsn: sys/sys/file.h 8.3
SCCS-vsn: sys/sys/namei.h 8.5
SCCS-vsn: sys/sys/mman.h 8.2
SCCS-vsn: sys/sys/time.h 8.3
SCCS-vsn: sys/sys/cdefs.h 8.8
SCCS-vsn: sys/sys/un.h 8.2
SCCS-vsn: sys/sys/resource.h 8.4
SCCS-vsn: sys/sys/resourcevar.h 8.4
SCCS-vsn: sys/sys/signalvar.h 8.5
SCCS-vsn: sys/sys/conf.h 8.5
SCCS-vsn: sys/sys/tty.h 8.7
SCCS-vsn: sys/sys/stat.h 8.10
SCCS-vsn: sys/sys/vnode.h 8.12
SCCS-vsn: sys/sys/systm.h 8.5
SCCS-vsn: sys/sys/buf.h 8.8
SCCS-vsn: sys/sys/mount.h 8.16
SCCS-vsn: sys/sys/ioccom.h 8.3

+++ usr/src/sys/sys/ioccom.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)ioccom.h	8.2 (Berkeley) %G%
+ *	@(#)ioccom.h	8.3 (Berkeley) %G%
  */
 
 #ifndef	_SYS_IOCCOM_H_
@@ -20,12 +20,17 @@
 #define	IOCBASECMD(x)	((x) & ~(IOCPARM_MASK << 16))
 #define	IOCGROUP(x)	(((x) >> 8) & 0xff)
 
-#define	IOCPARM_MAX	NBPG		/* max size of ioctl, mult. of NBPG */
-#define	IOC_VOID	0x20000000	/* no parameters */
-#define	IOC_OUT		0x40000000	/* copy out parameters */
-#define	IOC_IN		0x80000000	/* copy in parameters */
+#define	IOCPARM_MAX	NBPG	/* max size of ioctl args, mult. of NBPG */
+				/* no parameters */
+#define	IOC_VOID	(unsigned long)0x20000000
+				/* copy parameters out */
+#define	IOC_OUT		(unsigned long)0x40000000
+				/* copy parameters in */
+#define	IOC_IN		(unsigned long)0x80000000
+				/* copy paramters in and out */
 #define	IOC_INOUT	(IOC_IN|IOC_OUT)
-#define	IOC_DIRMASK	0xe0000000	/* mask for IN/OUT/VOID */
+				/* mask for IN/OUT/VOID */
+#define	IOC_DIRMASK	(unsigned long)0xe0000000
 
 #define	_IOC(inout,group,num,len) \
 	(inout | ((len & IOCPARM_MASK) << 16) | ((group) << 8) | (num))
